+++ R/paletteer_d.R
@@ -1,10 +1,9 @@
 #' Get discrete palette by package and name
 #'
 #' Available package/palette combinations are available in the data.frame
-#' \code{\link[paletteer]{palettes_d_names}}. Both `package` and `palette`
-#' can be supplied as symbols or strings.
+#' \code{\link[paletteer]{palettes_d_names}}.
 #'
-#' @param palette Name of palette as string or symbol. Must be on the form
+#' @param palette Name of palette as a string. Must be on the form
 #' packagename::palettename.
 #' @param n Number of colors desired. If omitted, returns complete palette.
 #' @param direction Either `1` or `-1`. If `-1` the palette will be reversed.
@@ -12,7 +11,6 @@
 #'   "continuous" is picked. Defaults to "discrete".
 #' @return A vector of colors.
 #' @examples
-#' paletteer_d(`nord::frost`)
 #' paletteer_d("nord::frost")
 #' paletteer_d("wesanderson::Royal1", 3)
 #' paletteer_d("Redmonder::dPBIPuOr", 14, type = "continuous")
@@ -26,12 +24,12 @@ paletteer_d <- function (palette, n, direction = 1,
 
   type <- match.arg(type)
 
-  palette <- rlang::quo_name(rlang::enquo(palette))
   palette <- unlist(strsplit(palette, "::"))
 
   pal <- paletteer::palettes_d[[palette]]
   if (is.null(pal))
-    stop("Palette not found. Make sure the palette name are spelled correct.")
+    stop('Palette not found. Make sure both package and palette ',
+         'name are spelled correct in the format "package::palette"')
   if (missing(n)) {
     n <- length(pal)
   }