+++ tests/testthat/test-paletteer_dynamic.R
@@ -18,11 +18,6 @@ test_that("paletteer_dynamic throws an error when n is specified too high", {
   expect_error(paletteer_dynamic("ggthemes_solarized::green", 100))
 })
 
-test_that("paletteer_dynamic works when package and name are provided as symbols", {
-  expect_length(paletteer_dynamic("ggthemes_solarized::green", 4), 4)
-  expect_length(paletteer_dynamic(`ggthemes_solarized::green`, 4), 4)
-})
-
 test_that("direction works correctly in paletteer_dynamic", {
   expect_equal(paletteer_dynamic("ggthemes_solarized::green", 3, direction = 1),
                c("#859900", "#dc322f", "#268bd2"))