+++ NEWS.md
@@ -1,6 +1,7 @@
 # paletteer (development version)
 
 * All front facing functions now take a single `palette` argument of the format `package::palette` instead of the  previous two arguments package and palette.
+* `palette` argument must be supplied as a string.
 
 # paletteer 0.2.1